fix(HomePage): render databases when whoami returns 403
